otlp builder pattern
