+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/.publicApi/net462/PublicAPI.Unshipped.txt
@@ -1,0 +1,1 @@
+~OpenTelemetry.Exporter.OtlpExporterOptions.endpoint -> System.Uri
