+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/.publicApi/net6.0/PublicAPI.Unshipped.txt
@@ -1,0 +1,1 @@
+~OpenTelemetry.Exporter.OtlpExporterOptions.endpoint -> System.Uri
