+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/.publicApi/netstandard2.1/PublicAPI.Unshipped.txt
@@ -1,0 +1,1 @@
+~OpenTelemetry.Exporter.OtlpExporterOptions.endpoint -> System.Uri
